[colorspaces.base] refacto immutable dataclass
part 1 just freezing the dataclass
This was decided mainly because of the relation between the matrix on RgbColorspaces and the primaries/whitepoint. If one one change the primaries of the child Gamut class, you would need to recompute the matrix of the parent RgbColorspace.

This also allow to cache properties/methods and remove previous "repr caching hacks"